+++ src/main/java/pkg1/library/BookRepo.java
@@ -1,3 +1,4 @@
+
 package pkg1.library;
 
 import java.util.List;
@@ -6,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface BookRepo extends JpaRepository<BookEntity,Integer>{
+public interface BookRepo extends JpaRepository<BookEntity,Long>{
 	@Query("SELECT b FROM BookEntity b WHERE LOWER(b.title) LIKE LOWER(CONCAT('%', :keyword, '%')) OR LOWER(b.author) LIKE LOWER(CONCAT('%', :keyword, '%'))")
 	List<BookEntity> searchBooks(@Param("keyword") String keyword);
