+++ src/main/java/pkg1/library/MemberRepo.java
@@ -2,6 +2,6 @@ package pkg1.library;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepo extends JpaRepository<MemberEntity, Integer>{
+public interface MemberRepo extends JpaRepository<MemberEntity, Long>{
 
 }
