+++ src/main/java/pkg1/library/StaffRepo.java
@@ -2,6 +2,6 @@ package pkg1.library;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffRepo extends JpaRepository<StaffEntity, Integer> {
+public interface StaffRepo extends JpaRepository<StaffEntity, Long> {
 
 }
