+++ src/main/java/pkg1/library/TransactionsRepo.java
@@ -2,6 +2,6 @@ package pkg1.library;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionsRepo extends JpaRepository<TransactionsEntity, Integer>{
+public interface TransactionsRepo extends JpaRepository<TransactionsEntity, Long>{
 
 }
